Audio Manager bug fix
For non looping audio effects, the FadeOut() function now doesnt do anything.

Instead the FadeIn() calls  PlayClip() which creates a new audio component with the same preferences, plays the audio until it ends, and deletes the component once done.

This is done in case we have the same audio clip repeating multiple times in small period of time.